feat: Implement monitoring stack with OpenAPI processing, dashboard generation, and KQL query functions
