+++ packages/cdktf-monitoring-stack/src/uri-utils.ts
@@ -1,0 +1,10 @@
+/**
+ * URI transformation utilities
+ */
+
+/**
+ * Translate path parameters of a URI to a generic version thanks to regex
+ */
+export function uriToRegex(value: string): string {
+  return String(value).replace(/{[^/]+}/g, "[^/]+") + "$";
+}
